requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
geoalchemy2==0.14.3
alembic==1.13.1

# Geospatial
rasterio==1.3.9
rio-cogeo==5.1.1
pyproj==3.6.1
shapely==2.0.2
geojson==3.1.0
pystac-client==0.7.5
planetary-computer==1.0.0

# Image processing
numpy==1.26.3
scipy==1.12.0
scikit-image==0.22.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0

# ML/Object Detection
ultralytics==8.1.9
torch==2.2.0
torchvision==0.17.0

# Storage
boto3==1.34.34
minio==7.2.3

# Task queue
celery[redis]==5.3.6
redis==5.0.1

# Tile serving
titiler.core==0.16.1
titiler.mosaic==0.16.1

# HTTP client
httpx==0.26.0
aiofiles==23.2.1

# Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Utilities
orjson==3.9.12
tenacity==8.2.3
structlog==24.1.0
